refactor: extract wezterm-surface from termwiz
