+++ wezterm-dynamic/derive/src/attr.rs
@@ -141,7 +141,7 @@ impl<'a> FieldInfo<'a> {
                     quote!(
                         #ident: match obj.get_by_str(#name) {
                             Some(v) => {
-                                use std::convert::TryFrom;
+                                use core::convert::TryFrom;
                                 #check_deprecated
                                 let target = <#try_from>::from_dynamic(v, options)
                                     .map_err(|source| source.field_context(
@@ -168,7 +168,7 @@ impl<'a> FieldInfo<'a> {
                     quote!(
                         #ident: match obj.get_by_str(&#name) {
                             Some(v) => {
-                                use std::convert::TryFrom;
+                                use core::convert::TryFrom;
                                 #check_deprecated
                                 let target = <#try_from>::from_dynamic(v, options)
                                     .map_err(|source| source.field_context(
@@ -194,7 +194,7 @@ impl<'a> FieldInfo<'a> {
                 DefValue::None => {
                     quote!(
                         #ident: {
-                            use std::convert::TryFrom;
+                            use core::convert::TryFrom;
                             let target = <#try_from>::from_dynamic(obj.get_by_str(#name).map(|v| {
                                 #check_deprecated
                                 v
